Removed sparseProbVec, which is now its own package. Added vech and xpnd_tri to general tools.

+++ src/conjugateUpdates.jl
@@ -16,12 +16,12 @@ export post_norm_mean, post_norm_var, rpost_normlm_beta1;
 
   Accepts the sample mean and returns a normal distribution.
 """
-function post_norm_mean(ȳ::Float64, n::Int, σ2::Float64, μ0::Float64, v0::Float64)
+function post_norm_mean(ȳ::Float, n::Int, σ2::Float, μ0::Float, v0::Float)
   σ2 > 0.0 || throw(ArgumentError("σ must be positive."))
   v0 > 0.0 || throw(ArgumentError("v0 must be positive."))
   n > 0.0 || throw(ArgumentError("n must be positive."))
 
-  # n = convert(Float64, n)
+  # n = convert(Float, n)
 
   v1 = 1.0 / (1.0/v0 + n/σ2)
   sd1 = sqrt(v1)
@@ -43,11 +43,11 @@ end
   `a0` is the shape parameter of the inverse-gamma prior.
   `b0` is the scale parameter of the inverse-gamma prior.
 """
-function post_norm_var(ss::Float64, n::Int64, a0::Float64, b0::Float64)
+function post_norm_var(ss::Float, n::Float, a0::Float, b0::Float)
   a0 > 0.0 || throw(ArgumentError("a0 must be positive."))
   b0 > 0.0 || throw(ArgumentError("b0 must be positive."))
 
-  # n = convert(Float64, n)
+  # n = convert(Float, n)
   a1 = a0 + 0.5*n
   b1 = b0 + 0.5*ss
 
@@ -65,9 +65,9 @@ end
 
   Assumes independent normals prior with mean `0.0` and variance `τ2`.
 """
-function rpost_normlm_beta1(y::Vector{Float64},
-  X::Matrix{Float64},
-  σ2::Float64, τ2::Float64, β0::Float64=0.0)
+function rpost_normlm_beta1(y::Vector{Float},
+  X::Matrix{Float},
+  σ2::Float, τ2::Float, β0::Float=0.0)
   # n > 1 and p > 1 case
 
   σ2 > 0.0 || throw(ArgumentError("σ2 must be positive."))
@@ -88,8 +88,8 @@ function rpost_normlm_beta1(y::Vector{Float64},
   z = randn(p)
   β = U \ z + μ
 end
-function rpost_normlm_beta1(y::Float64, X::Vector{Float64},
-  σ2::Float64, τ2::Float64, β0::Float64=0.0)
+function rpost_normlm_beta1(y::Float, X::Vector{Float},
+  σ2::Float, τ2::Float, β0::Float=0.0)
   # n = 1 and p > 1 case
 
   σ2 > 0.0 || throw(ArgumentError("σ2 must be positive."))
@@ -109,9 +109,9 @@ function rpost_normlm_beta1(y::Float64, X::Vector{Float64},
   z = randn(p)
   β = U \ z + μ
 end
-function rpost_normlm_beta1(y::Vector{Float64},
-  X::Vector{Float64},
-  σ2::Float64, τ2::Float64, β0::Float64=0.0)
+function rpost_normlm_beta1(y::Vector{Float},
+  X::Vector{Float},
+  σ2::Float, τ2::Float, β0::Float=0.0)
   # n > 1 and p = 1 case
 
   σ2 > 0.0 || throw(ArgumentError("σ2 must be positive."))
@@ -126,8 +126,8 @@ function rpost_normlm_beta1(y::Vector{Float64},
   z = randn(1)
   β = z / sqrt(A) .+ μ
 end
-function rpost_normlm_beta1(y::Float64, X::Float64,
-  σ2::Float64, τ2::Float64, β0::Float64=0.0)
+function rpost_normlm_beta1(y::Float, X::Float,
+  σ2::Float, τ2::Float, β0::Float=0.0)
   # n = 1 and p = 1 case
 
   σ2 > 0.0 || throw(ArgumentError("σ2 must be positive."))
